requirements.txt
pyusb
evdev
obs-websocket-py
pillow